pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["kaori*"]

[project]
name = "kaori"
version = "0.2.0"
description = "Kaori — Personal AI-powered life management"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "pillow>=11.0",
    "pillow-heif>=1.0",
    "python-multipart>=0.0.18",
    "aiosqlite>=0.21",
]

[project.optional-dependencies]
llm-api = [
    "anthropic>=0.43",
]
dev = [
    "pytest>=8.0",
    "httpx>=0.28",
    "pytest-asyncio>=0.25",
]
